Z-Levels
Z-levels are now calculated better, a point on the player character is compared to the collision bounds of nearby entities to see if the character is in front of or behind said entities. Does not work well on edges, the point should be made into a horizontal line instead. Will also run into problems when the character is behind an object with a lower z-level than an object they're in front of.